feat: enhance message port connection between 'background.js' and 'timetable.js'.

+++ extension/scripts/background.js
@@ -155,16 +155,24 @@ chrome.runtime.onMessage.addListener((message, sender, sendResponse) => {
             console.log("Pure backtracking result: ", pureComb);
             console.log("Backtracking with daytime conflict + seats availability: ", prunedComb);
 
-            // Passing pruned combination to 'result.html'
+            // Passing pruned combination to 'timetable.html'
             chrome.tabs.create({ url: chrome.runtime.getURL("./timetable/timetable.html") }, () => {
-                chrome.runtime.sendMessage({ action: "passDataset", dataset: prunedComb }, () => {
-                    console.log("Pruned dataset sent to result.html");
+                chrome.runtime.onMessage.addListener((message, sender, sendResponse) => {
+                    if(message.action === "timetablejsInjected") {
+                        chrome.runtime.sendMessage({ action: "passDataset", dataset: prunedComb }, (response) => {
+                            console.log("Pruned dataset sent to timetable.html");
+
+                            // Clear chrome storage
+                            chrome.storage.local.clear();
+                            console.log("Dataset cleared from chrome storage.");
+
+                            console.log(response.status);
+                        });
+                    }
+                    return true; // keep message port open for receiving message
                 });
             });
 
-            // Clear chrome storage
-            chrome.storage.local.clear();
-            console.log("Dataset cleared from chrome storage.");
         });
 
         // Pure Backtracking
